requirements.txt
python-dotenv>=1.0

fastapi[standard]>=0.110,<1.0

httpx
